requirements.txt
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.9.2
natsort==8.4.0
numpy==2.0.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
QtPy==2.4.1
scipy==1.13.1
six==1.16.0
tqdm==4.66.5
ujson==5.10.0
zipp==3.20.2
nodeeditor==0.9.15
